Add psycopg2 and psycopg2-binary as extra dependency options
**Sumary:**
Currently we can't install the `bitdotio` package on Deepnote since the
`psycopg2` dependency has some build prerequisites that cannot be met.

This PR removes the `psycopg2` dependency from the default installation
of the `bitdotio` package and uses the `extras_require` parameter to allow
the user to choose which dependency they would like to install, either
`psycopg2` if they have Postgres already installed or `psycopg2-binary`
which doesn't depend on Postgres.